+++ JustDX12/ModelLoading/TextureLoader.cpp
@@ -51,14 +51,14 @@ void TextureLoader::loadTexture(DX12Texture* tex) {
 	assert(texMetaData.dimension != D3D12_RESOURCE_DIMENSION_TEXTURE1D);
 
 	tex->MetaData.Format = texMetaData.format;
-	tex->MetaData.Width = texMetaData.width;
-	tex->MetaData.Height = texMetaData.height;
+	tex->MetaData.Width = (UINT)texMetaData.width;
+	tex->MetaData.Height = (UINT)texMetaData.height;
 	if (texMetaData.format == DXGI_FORMAT_BC5_UNORM) {
-		tex->MetaData.Width = DivRoundUp(texMetaData.width, 4) * 4;
-		tex->MetaData.Height = DivRoundUp(texMetaData.height, 4) * 4;
+		tex->MetaData.Width = (UINT)DivRoundUp(texMetaData.width, 4) * 4;
+		tex->MetaData.Height = (UINT)DivRoundUp(texMetaData.height, 4) * 4;
 	}
 	tex->MetaData.Flags = D3D12_RESOURCE_FLAG_NONE;
-	tex->MetaData.DepthOrArraySize = texMetaData.arraySize;
+	tex->MetaData.DepthOrArraySize = (UINT16)texMetaData.arraySize;
 	tex->MetaData.MipLevels = (UINT16)texMetaData.mipLevels;
 	tex->MetaData.SampleDesc.Count = 1;
 	tex->MetaData.SampleDesc.Quality = 0;
